aggiornamento controller

+++ GiftifyBackend/Giftify/src/main/java/it/unicam/cs/Giftify/Controller/CommunityController.java
@@ -2,6 +2,8 @@ package it.unicam.cs.Giftify.Controller;
 
 import it.unicam.cs.Giftify.Model.Entity.Account;
 import it.unicam.cs.Giftify.Model.Entity.Community;
+import it.unicam.cs.Giftify.Model.Entity.Role;
+import it.unicam.cs.Giftify.Model.Services.AccountService;
 import it.unicam.cs.Giftify.Model.Services.CommunityService;
 import it.unicam.cs.Giftify.Model.Util.AccessCodeGeneretor;
 import it.unicam.cs.Giftify.Model.Util.DTOClasses.CommunityCreateDTO;
@@ -9,55 +11,68 @@ import it.unicam.cs.Giftify.Model.Util.DTOClasses.CommunityUpdateDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 
-@RestController
+@RestController("/community")
 public class CommunityController {
 
     @Autowired
     private CommunityService communityService;
 
     @Autowired
+    private AccountService accountService;
+
+    @Autowired
     private AccessCodeGeneretor accessCodeGeneretor;
 
     @PostMapping("/create")
     public ResponseEntity<String> createCommunity(@RequestBody CommunityCreateDTO communityDto) {
-        try {
-            Account admin = communityService.findAccountById(communityDto.getAdminId());
-            if (admin == null) {
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Admin non trovato.");
-            }
-            communityService.createCommunity(
-                    accessCodeGeneretor,
-                    admin,
-                    communityDto.getCommunityName(),
-                    communityDto.getNote(),
-                    communityDto.getBudget(),
-                    communityDto.getDeadline()
-            );
-            return ResponseEntity.ok("Community creata con successo.");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(e.getMessage());
-        }
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account admin = (Account) authentication.getPrincipal();
+        admin = accountService.getAccountById(admin.getId());
+        communityService.createCommunity(
+                accessCodeGeneretor,
+                admin,
+                communityDto.getCommunityName(),
+                communityDto.getNote(),
+                communityDto.getBudget(),
+                communityDto.getDeadline()
+        );
+        return ResponseEntity.ok("Community creata con successo.");
     }
 
-    @PreAuthorize("hasRole('ROLE_ADMIN_' + #communityId)")
-    @DeleteMapping("/removeUser")
+
+    @DeleteMapping("/removeUser/{id}")
     public ResponseEntity<Void> removeUserFromCommunity(@PathVariable Long communityId, @RequestBody Account user) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account admin = (Account) authentication.getPrincipal();
+        admin = accountService.getAccountById(admin.getId());
         Community community = communityService.getCommunityById(communityId);
+        if (!admin.getRoleForCommunity(community).equals(Role.ADMIN)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
+        if (admin == null) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        }
         communityService.removeUserFromCommunity(user, community);
         return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
-    @PreAuthorize("hasRole('ROLE_ADMIN_' + #communityId)")
+
     @PostMapping("/close/{id}")
     public ResponseEntity<String> closeCommunity(@PathVariable long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account admin = (Account) authentication.getPrincipal();
+        admin = accountService.getAccountById(admin.getId());
         Community community = communityService.getCommunityById(id);
+        if (!admin.getRoleForCommunity(community).equals(Role.ADMIN)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (community != null) {
             communityService.drawNames(community);
             return ResponseEntity.ok("Community chiusa con successo.");
@@ -67,7 +82,13 @@ public class CommunityController {
 
     @PostMapping("/delete/{id}")
     public ResponseEntity<String> deleteCommunity(@PathVariable long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account admin = (Account) authentication.getPrincipal();
+        admin = accountService.getAccountById(admin.getId());
         Community community = communityService.getCommunityById(id);
+        if (!admin.getRoleForCommunity(community).equals(Role.ADMIN)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (community != null) {
             communityService.deleteGroup(community);
             return ResponseEntity.ok("Community eliminata con successo.");
@@ -79,7 +100,13 @@ public class CommunityController {
     public ResponseEntity<String> updateCommunity(
             @PathVariable long id,
             @RequestBody CommunityUpdateDTO communityUpdateDto) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account admin = (Account) authentication.getPrincipal();
+        admin = accountService.getAccountById(admin.getId());
         Community community = communityService.getCommunityById(id);
+        if (!admin.getRoleForCommunity(community).equals(Role.ADMIN)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (community != null) {
             if (communityUpdateDto.getCommunityName() != null) {
                 community.setCommunityName(communityUpdateDto.getCommunityName());
@@ -96,24 +123,16 @@ public class CommunityController {
         return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Community non trovata.");
     }
 
-//    @GetMapping("/roles/{id}")
-//    public ResponseEntity<Map<Account, String>> getRoles(@PathVariable long id) {
-//        Community community = communityService.getCommunityById(id);
-//        if (community != null) {
-//            Map<Account, String> roles = new HashMap<>();
-//            roles.put(community.getAdmin(), "Admin");
-//            for (Account user : community.getUserList()) {
-//                roles.putIfAbsent(user, "Member");
-//            }
-//            return ResponseEntity.ok(roles);
-//        }
-//        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(null);
-//    }
 
-    @PreAuthorize("@communityService.isUserParticipantOfCommunity(#id, principal)")
     @GetMapping("/participants/{id}")
     public ResponseEntity<List<Account>> getParticipants(@PathVariable long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account account = (Account) authentication.getPrincipal();
+        account = accountService.getAccountById(account.getId());
         Community community = communityService.getCommunityById(id);
+        if (account.getRoleForCommunity(community) == null) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (community != null) {
             return ResponseEntity.ok(community.getUserList());
         }
@@ -122,7 +141,13 @@ public class CommunityController {
 
     @GetMapping("/info/{id}")
     public ResponseEntity<Community> getGeneralInfo(@PathVariable long id) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account account = (Account) authentication.getPrincipal();
+        account = accountService.getAccountById(account.getId());
         Community community = communityService.getCommunityById(id);
+        if (account.getRoleForCommunity(community) == null) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (community != null) {
             return ResponseEntity.ok(community);
         }
@@ -131,7 +156,13 @@ public class CommunityController {
 
     @GetMapping("/draw/{id}")
     public ResponseEntity<Account> viewDrawnName(@PathVariable long id, @RequestParam long userId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Account admin = (Account) authentication.getPrincipal();
+        admin = accountService.getAccountById(admin.getId());
         Community community = communityService.getCommunityById(id);
+        if (!admin.getRoleForCommunity(community).equals(Role.ADMIN)) {
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN);
+        }
         if (community != null) {
             Account user = community.getUserList().stream()
                     .filter(u -> u.getId().equals(userId))
